[Rust]: Fix `BlockchainType` deserialization
* Remove `CoinType` arguments from FFI functions

+++ rust/wallet_core_rs/tests/ethereum_abi.rs
@@ -57,7 +57,7 @@ fn test_ethereum_abi_decode_contract_call() {
     let input_data = TWDataHelper::create(serialize(&input).unwrap());
 
     let output_data = TWDataHelper::wrap(unsafe {
-        tw_ethereum_abi_decode_contract_call(CoinType::Ethereum, input_data.ptr())
+        tw_ethereum_abi_decode_contract_call(CoinType::Ethereum as u32, input_data.ptr())
     })
     .to_vec()
     .expect("!tw_ethereum_abi_decode_contract_call returned nullptr");
@@ -97,7 +97,7 @@ fn test_ethereum_abi_decode_params() {
     let input_data = TWDataHelper::create(serialize(&input).unwrap());
 
     let output_data = TWDataHelper::wrap(unsafe {
-        tw_ethereum_abi_decode_params(CoinType::Ethereum, input_data.ptr())
+        tw_ethereum_abi_decode_params(CoinType::Ethereum as u32, input_data.ptr())
     })
     .to_vec()
     .expect("!tw_ethereum_abi_decode_params returned nullptr");
@@ -134,7 +134,7 @@ fn test_ethereum_abi_function_get_signature() {
     let input_data = TWDataHelper::create(serialize(&input).unwrap());
 
     let actual = TWStringHelper::wrap(unsafe {
-        tw_ethereum_abi_function_get_signature(CoinType::Ethereum, input_data.ptr())
+        tw_ethereum_abi_function_get_signature(CoinType::Ethereum as u32, input_data.ptr())
     })
     .to_string()
     .expect("!tw_ethereum_abi_function_get_signature returned nullptr");
@@ -155,7 +155,7 @@ fn test_ethereum_abi_encode_function() {
     let input_data = TWDataHelper::create(serialize(&input).unwrap());
 
     let output_data = TWDataHelper::wrap(unsafe {
-        tw_ethereum_abi_encode_function(CoinType::Ethereum, input_data.ptr())
+        tw_ethereum_abi_encode_function(CoinType::Ethereum as u32, input_data.ptr())
     })
     .to_vec()
     .expect("!tw_ethereum_abi_encode_function returned nullptr");
@@ -181,7 +181,7 @@ fn test_ethereum_abi_decode_value() {
     let input_data = TWDataHelper::create(serialize(&input).unwrap());
 
     let output_data = TWDataHelper::wrap(unsafe {
-        tw_ethereum_abi_decode_value(CoinType::Ethereum, input_data.ptr())
+        tw_ethereum_abi_decode_value(CoinType::Ethereum as u32, input_data.ptr())
     })
     .to_vec()
     .expect("!tw_ethereum_abi_decode_value returned nullptr");
